Dispatch to other elements
Here it also becomes clear why the ProcessSettled should be a proxy
class and not the document itself. Running all of this on the UI thread
just slows the page while there is computation ongoing. Realistically we
want to put all the real work into webworkers asap. There is no standard
way to invoke a function with 'gas' (as wasmtime calls it), i.e. yield
by VM breakpoints periodically.